requirements.txt
python-twitter==3.5
praw==6.5.1
Pillow==5.3.0